refactor: split frontend common ui bundle v3
